# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API behavior, compatibility rules, and endpoint contract requirements.",
    "version": "0.5.0",
    "description": "- response style: JSON (Pydantic response models); binary endpoints return file streams (FileResponse / StreamingResponse)."
  },
  "paths": {
...
    "/jobs": {
      "post": {
        "summary": "POST /jobs",
        "responses": {
          "201": {
            "description": "Contract response: JobCreateResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "422": {
            "description": "Error response (see contract error format)"
          }
        },
        "requestBody": {
          "description": "Contract request: multipart/form-data (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "400, 422"
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "GET /jobs/{job_id}",
        "responses": {
          "200": {
            "description": "Contract response: JobStatus",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "404": {
            "description": "Error response (see contract error format)"
          }
        },
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-cdd-errors": "404"
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "summary": "POST /jobs/{job_id}/cancel",
        "responses": {
          "201": {
            "description": "Contract response: -"
          },
          "404": {
            "description": "Error response (see contract error format)"
          }
        },
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-cdd-response-contract": "-",
        "x-cdd-errors": "404"
      }
    },
    "/jobs/{job_id}/download": {
      "get": {
        "summary": "GET /jobs/{job_id}/download",
        "responses": {
          "200": {
            "description": "Contract response: file stream"
          },
          "404": {
            "description": "Error response (see contract error format)"
          }
        },
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-cdd-response-contract": "file stream",
        "x-cdd-errors": "404"
      }
    },
...
    "/terms/import": {
      "post": {
        "summary": "POST /terms/import",
        "responses": {
          "201": {
            "description": "Contract response: TermImportResult",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermImportResult"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "422": {
            "description": "Error response (see contract error format)"
          }
        },
        "requestBody": {
          "description": "Contract request: multipart/form-data (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "400, 422"
      }
    },
...
    "/terms/approve": {
      "post": {
        "summary": "POST /terms/approve",
        "responses": {
          "201": {
            "description": "Contract response: -"
          },
          "404": {
            "description": "Error response (see contract error format)"
          }
        },
        "requestBody": {
          "description": "Contract request: TermApproveRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermApproveRequest"
              }
            }
          }
        },
        "x-cdd-response-contract": "-",
        "x-cdd-errors": "404"
      }
    },
    "/terms/reject": {
      "post": {
        "summary": "POST /terms/reject",
        "responses": {
          "201": {
            "description": "Contract response: -"
          },
          "404": {
            "description": "Error response (see contract error format)"
          }
        },
        "requestBody": {
          "description": "Contract request: TermRejectRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermRejectRequest"
              }
            }
          }
        },
        "x-cdd-response-contract": "-",
        "x-cdd-errors": "404"
      }
    },
    "/terms/flag-needs-review": {
      "post": {
        "summary": "POST /terms/flag-needs-review",
        "responses": {
          "201": {
            "description": "Contract response: -"
          },
          "404": {
            "description": "Error response (see contract error format)"
          }
        },
        "requestBody": {
          "description": "Contract request: TermRejectRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermRejectRequest"
              }
            }
          }
        },
        "x-cdd-response-contract": "-",
        "x-cdd-errors": "404"
      }
    },
...
    "/terms/edit": {
      "patch": {
        "summary": "PATCH /terms/edit",
        "responses": {
          "200": {
            "description": "Contract response: -"
          },
          "404": {
            "description": "Error response (see contract error format)"
          }
        },
        "requestBody": {
          "description": "Contract request: TermEditRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermEditRequest"
              }
            }
          }
        },
        "x-cdd-response-contract": "-",
        "x-cdd-errors": "404"
      }
    },
    "/terms/wikidata/search": {
      "post": {
        "summary": "POST /terms/wikidata/search",
        "responses": {
          "201": {
            "description": "Contract response: WikidataSearchResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WikidataSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Error response (see contract error format)"
          }
        },
        "requestBody": {
          "description": "Contract request: WikidataSearchRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WikidataSearchRequest"
              }
            }
          }
        },
        "x-cdd-errors": "422"
      }
    },
    "/terms/wikidata/import": {
      "post": {
        "summary": "POST /terms/wikidata/import",
        "responses": {
          "201": {
            "description": "Contract response: -"
          },
          "422": {
            "description": "Error response (see contract error format)"
          }
        },
        "requestBody": {
          "description": "Contract request: WikidataImportRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WikidataImportRequest"
              }
            }
          }
        },
        "x-cdd-response-contract": "-",
        "x-cdd-errors": "422"
      }
    },
...
    "/jobs/{job_id}/quality": {
      "get": {
        "summary": "GET /jobs/{job_id}/quality",
        "responses": {
          "200": {
            "description": "Contract response: JobQualityResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobQualityResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error response (see contract error format)"
          }
        },
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-cdd-errors": "200 (status: available/pending/disabled/unavailable); 404 job not found"
      }
    }
  },
...
  "components": {
    "schemas": {
...
      "JobCreateResponse": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ]
      },
      "JobStatus": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "processed_files": {
            "type": "integer"
          },
          "total_files": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "output_ready": {
            "type": "boolean"
          },
          "current_file": {
            "type": "string"
          },
          "segments_done": {
            "type": "integer"
          },
          "segments_total": {
            "type": "integer"
          },
          "file_segments_done": {
            "type": "integer"
          },
          "file_segments_total": {
            "type": "integer"
          },
          "elapsed_seconds": {
            "type": "number"
          },
          "overall_progress": {
            "type": "number"
          },
          "segments_per_second": {
            "type": "number"
          },
          "eta_seconds": {
            "type": "number"
          },
          "term_summary": {
            "type": "string",
            "description": "per-language/domain term extraction counts as JSON-serialized map; null when mode != extraction_only"
          },
          "provider": {
            "type": "string",
            "description": "provider ID that successfully processed this job (e.g. panjit, deepseek, ollama-local); null if not yet determined or Ollama-only legacy job"
          }
        },
        "required": [
          "job_id",
          "status",
          "processed_files",
          "total_files",
          "output_ready",
          "elapsed_seconds",
          "overall_progress",
          "segments_per_second"
        ]
      },
...
      "TermImportResult": {
        "type": "object",
        "properties": {
          "inserted": {
            "type": "integer"
          },
          "skipped": {
            "type": "integer"
          },
          "overwritten": {
            "type": "integer"
          }
        },
        "required": [
          "inserted",
          "skipped",
          "overwritten"
        ]
      },
      "TermApproveRequest": {
        "type": "object",
        "properties": {
          "source_text": {
            "type": "string"
          },
          "target_lang": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          }
        },
        "required": [
          "source_text",
          "target_lang",
          "domain"
        ]
      },
      "TermRejectRequest": {
        "type": "object",
        "properties": {
          "source_text": {
            "type": "string"
          },
          "target_lang": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          }
        },
        "required": [
          "source_text",
          "target_lang",
          "domain"
        ]
      },
      "TermEditRequest": {
        "type": "object",
        "properties": {
          "source_text": {
            "type": "string"
          },
          "target_lang": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "target_text": {
            "type": "string"
          },
          "confidence": {
            "type": "number"
          }
        },
        "required": [
          "source_text",
          "target_lang",
          "domain",
          "target_text"
        ]
      },
      "WikidataSearchRequest": {
        "type": "object",
        "properties": {
          "term": {
            "type": "string"
          },
          "source_lang": {
            "type": "string",
            "description": "default Chinese"
          },
          "target_langs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "default English"
          },
          "domain": {
            "type": "string",
            "description": "default general"
          }
        },
        "required": [
          "term"
        ]
      },
...
      "WikidataSearchResponse": {
        "type": "object",
        "properties": {
          "term": {
            "type": "string"
          },
          "candidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WikidataCandidate"
            }
          }
        },
        "required": [
          "term",
          "candidates"
        ]
      },
      "WikidataImportRequest": {
        "type": "object",
        "properties": {
          "source_text": {
            "type": "string"
          },
          "target_text": {
            "type": "string"
          },
          "source_lang": {
            "type": "string"
          },
          "target_lang": {
            "type": "string"
          },
          "domain": {
            "type": "string",
            "description": "default general"
          },
          "entity_id": {
            "type": "string"
          }
        },
        "required": [
          "source_text",
          "target_text",
          "source_lang",
          "target_lang"
        ]
      },
...
      "JobQualityResponse": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string",
            "description": "job identifier"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "pending",
              "disabled",
              "unavailable"
            ],
            "description": "available — scores ready; pending — translation not yet complete or scoring still running; disabled — QE_ENABLED=false; unavailable — model load failed or scoring failed for this job"
          },
          "scores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockQualityScore"
            },
            "description": "present and non-empty when status = available; omitted or empty array when status is any other value"
          }
        },
        "required": [
          "job_id",
          "status"
        ]
      }
    }
  }
}